+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [feat: (#38) index operator](https://github.com/amp-lang/amp/issues/37)
 - [feat: (#38) subslice operator](https://github.com/amp-lang/amp/issues/37)
 - feat: external names for functions
+- [feat: (#39) variadic function arguments](https://github.com/amp-lang/amp/issues/39)
 
 ### Library
 - feat: implement standard library, I guess